requirements.txt
aiogram==3.20.0
python-docx==1.2.0
PyPDF2==3.0.1
aiofiles==24.1.0

